requirements.txt
django
psycopg2-binary
djangorestframework
django-filter
tensorflow
numpy
pandas
scipy
sklearn
alpha_vantage
keras